Split ValueWithUnits tests into a new file
