+++ src/test/unittests.sh
@@ -8,6 +8,7 @@ trap 'err=1' ERR
 python3 ../hammer-vlsi/test.py
 python3 ../hammer-vlsi/cli_driver_test.py
 python3 ../hammer-vlsi/utils_test.py
+python3 ../hammer-vlsi/units_test.py
 python3 ../hammer-vlsi/verilog_utils_test.py
 python3 ../hammer_config_test/test.py
